+++ R/utils.R
@@ -28,6 +28,21 @@ skip_if_no_mistnet <- function() {
   testthat::skip("No MistNet")
 }
 
+#' Skip test if vol2birdR not installed
+#'
+#' Some functions require suggested package vol2birdR to be installed.
+#' This helper function allows to skip a test if vol2birdR is not available, e.g. when running in CI.
+#' Inspired by <https://testthat.r-lib.org/articles/skipping.html#helpers>.
+#' @return Invisibly returns TRUE if vol2birdR is installed, otherwise skips the test with
+#' a message "Package vol2birdR not installed".
+#' @keywords internal
+skip_if_no_vol2birdR <- function() {
+  if (rlang::is_installed("vol2birdR")) {
+    return(invisible(TRUE))
+  }
+  testthat::skip("Package vol2birdR not installed")
+}
+
 #' Check if radar codes are exactly 5 characters
 #'
 #' @param radars character vector. Radar codes to check, e.g. `c("bejab",