+++ NEWS.md
@@ -1,3 +1,9 @@
+# bioRad 0.7.1.9000
+
+* changed default aloft bucket to aloftdata (#622)
+
+* skip tests for `calculate_vp()` when vol2birdR package is not installed (#624)
+
 # bioRad 0.7.1
 
 Rebuilds documentation with examples formatted as per CRAN requirements.